C/Linux: Mark the load of non-value-return atomics as "noreturn" (#4)
On some architectures, there exists atomic memory operations which will
be run directly at memory controller after issued by a CPU. Such an
operation doesn't return a value to a register, therefore the load part
is not treated as a load by barriers. In order to differ this kind of
operations, we mark the load part of a non-value-return atomic as
"noreturn" annotation. The memory model's cat file can add special rules
to it afterwards.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>